HYRAX-667, clean up code:remove old commented out code, use range-for, const, auto and default/override.

+++ modules/hdf4_handler/HDFEOS2ArrayMissField.cc
@@ -69,7 +69,7 @@ bool HDFEOS2ArrayMissGeoField::read()
 int
 HDFEOS2ArrayMissGeoField::format_constraint (int *offset, int *step, int *count)
 {
-    long nels = 1;
+    int nels = 1;
     int id = 0;
 
     Dim_iter p = dim_begin ();
